Add files via upload
Explain how it works
-Doctor-patient relationship:
The doctor-patient relationship can be represented using a correlation. In this case, we can say that a doctor can diagnose and one or more patients, and a patient can be diagnosed by one or more doctors.
-Relationship between nurse and patient:
The relationship between the nurse and the patient can also be represented using an association relationship. In this case, we can say that a nurse can be assigned to take care of one or many patients with a disease, and the disease can be taken care of by one or several nurses.
-The relationship between the patient and the room:
The relationship between the patient and the room can be represented by using an aggregation relationship to reserve a room for the patient,
-The relationship between the drug and the patient: The relationship between the drug and the patient can be represented using the relationship of each patient to more than one drug and each drug to more than one patient,